# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Faithly API",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/card/{id}": {
            "get": {
                "tags": [
                    "card"
                ],
                "summary": "Returns a Sample API response",
                "description": "A sample greeting to test out the API",
                "operationId": "5ad5612b8bd76f932aadbc1b1e4245b2",
                "parameters": [
                    {
                        "name": "Authentication",
                        "in": "header",
                        "description": "Provides user authentication bearer token",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "Success List Card",
                                    "properties": {
                                        "meta": {
                                            "description": "",
                                            "properties": {
                                                "response_code": {
                                                    "description": "",
                                                    "type": "integer",
                                                    "example": 200,
                                                    "nullable": false
                                                },
                                                "response_message": {
                                                    "description": "",
                                                    "type": "string",
                                                    "example": "Success",
                                                    "nullable": false
                                                }
                                            },
                                            "type": "object",
                                            "nullable": false
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CardPayment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Servier Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "Success List Card",
                                    "properties": {
                                        "meta": {
                                            "description": "",
                                            "properties": {
                                                "response_code": {
                                                    "description": "",
                                                    "type": "integer",
                                                    "example": 500,
                                                    "nullable": false
                                                },
                                                "response_message": {
                                                    "description": "",
                                                    "type": "string",
                                                    "example": "",
                                                    "nullable": false
                                                },
                                                "response_debug_param": {
                                                    "description": "",
                                                    "type": "string",
                                                    "example": "internal Server Error",
                                                    "nullable": false
                                                }
                                            },
                                            "type": "object",
                                            "nullable": false
                                        },
                                        "data": {
                                            "description": "List of data",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "Success List Card",
                                    "properties": {
                                        "meta": {
                                            "description": "",
                                            "properties": {
                                                "response_code": {
                                                    "description": "",
                                                    "type": "integer",
                                                    "example": 404,
                                                    "nullable": false
                                                },
                                                "response_message": {
                                                    "description": "",
                                                    "type": "string",
                                                    "example": "Data Not Found",
                                                    "nullable": false
                                                },
                                                "response_debug_param": {
                                                    "description": "",
                                                    "type": "string",
                                                    "example": "",
                                                    "nullable": false
                                                }
                                            },
                                            "type": "object",
                                            "nullable": false
                                        },
                                        "data": {
                                            "description": "List of data",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CardPayment": {
                "properties": {
                    "id": {
                        "description": "Id of card payment",
                        "type": "integer",
                        "nullable": false
                    },
                    "user_detail_id": {
                        "description": "",
                        "type": "integer",
                        "nullable": false
                    },
                    "card_number": {
                        "description": "",
                        "type": "integer",
                        "nullable": false
                    },
                    "exp_date": {
                        "description": "",
                        "type": "string",
                        "nullable": false
                    },
                    "code_verification": {
                        "description": "",
                        "type": "string",
                        "nullable": false
                    },
                    "status_id": {
                        "description": "",
                        "type": "integer",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        }
    }
}